+++ HISTORY.md
@@ -1,5 +1,17 @@
 # Release History
 
+## 0.13.0rc1
+
+- feat: add asyncio-native AsyncLedgerClient
+
+## 0.12.2
+
+- refactor: dependencies - unlocked click, setuptools
+
+## 0.12.1
+
+- fix: missing packaging dependency
+
 ## 0.12.0
 
 - Upgrade to fetchai Cosmos-SDK v0.20.0 (canonical v0.53.7)